update Cart component to properly render via CartElement Component
